notifications/admin_panel/requirements.txt
Django==4.2
django-debug-toolbar==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
split_settings==1.0.0
uwsgi==2.0.20